Rename: routes => pageDirectory and PAGE_DATA => PAGE_DIRECTORY
setup naming convention:

If we think about the "type" of state change this action is related to, it is related to the `data` flowing into the app via a query. I would like to come up with naming convention for such case since we will have LOTS of such cases and I don't want us making arbitrary naming choices.

So let's say data in questions is: `dataName`

- ❓ SQL query: `{dataName}_QUERY`
- ☸️  JS transform/reducer: `{dataName}_TRANSFORM`
- 🚀  Redux action: `SET_{dataName}`

So based on the above lets take `PAGE_DIRECTORY ` as example of our `dataName`:

everything looks good but we just need to change `SET_ROUTES` to `SET_PAGE_DATA`. Note: routes is just one of the ways this `data` is used. It's also used to build the sidebar, so semantically it makes more sense to call it based on the `data`. Basically every component that gets this `data` decides what is does with that `data`.

- ❓ SQL query: `PAGE_DIRECTORY_QUERY`
- ☸️  JS transform/reducer: `PAGE_DIRECTORY_TRANSFORM`
- 🚀  Redux action: `SET_ PAGE_DIRECTORY`
ref #8